the design has been changed, delete 3 test_functions

+++ test_filemanager.py
@@ -1,19 +1,24 @@
 import Programms
-from Programms.use_functions import separator, put_cash, data, buy, traffic_money
+from Programms.use_functions import separator, put_cash, buy, traffic_money
 import datetime
 import os
 
-
 def test_separator():
     assert separator('*', 5) == '*****'
+    assert separator('Ж', 5) != ';;;;;'
 
 def test_mkdir():
-    os.mkdir('folder mk')
+    if not os.path.exists('folder mk'):
+        os.mkdir('folder mk')
     assert 'folder mk' in os.listdir()
-    os.rmdir('folder mk')
+    if os.path.exists('folder mk'):
+        os.rmdir('folder mk')
 
 def test_listdir():
-    assert os.listdir("C:/Users/Alex/PycharmProjects/PS/Lesson5")
+    path = "C:/Users/Alex/PycharmProjects/PS/Lesson5"
+    assert isinstance(os.listdir(path), list) == True
+    assert isinstance(os.listdir(path), dict) == False
+
 
 def test_path_exists():
     assert os.path.exists("C:/Users/Alex/PycharmProjects/PS/Lesson5/test_filemanager.py") == True
@@ -21,11 +26,7 @@ def test_path_exists():
 def test_path_isfile():
     assert os.path.isfile("C:/Users/Alex/PycharmProjects/PS/Lesson5/test_filemanager.py") == True
 
-def test_environ():
-    assert os.environ
 
-def test_getcwd():
-    assert os.getcwd()
